+++ src/libs/boards/board.cpp
@@ -6,7 +6,7 @@ PlanarBoard::PlanarBoard() {}
 PlanarBoard::~PlanarBoard() {}
 
 // No coordiante transform for base class.
-BoardPoint PlanarBoard::p(double x, double y, double z) {
+BoardPoint PlanarBoard::p(float x, float y, float z) {
   return BoardPoint(x, y, z);
 }
