+++ src/libs/boards/boardpoint.cpp
@@ -1,7 +1,7 @@
 // boardpoint.cpp //
 #include <ggp_robot/libs/boards/boardpoint.h>
 
-BoardPoint::BoardPoint(double x, double y, double z)
+BoardPoint::BoardPoint(float x, float y, float z)
   : x(x), y(y), z(z)
 {}
 
@@ -10,7 +10,7 @@ std::ostream& operator<<(std::ostream& out, const BoardPoint& p) {
   return out << "BoardPoint(" << p.x << "|" << p.y << "|" << p.z << ")";
 }
 
-BoardPoint::operator cv::Point3d() { return cv::Point3d(x,y,z); }
-BoardPoint::operator cv::Point2d() { return cv::Point2d(x,y); }
+BoardPoint::operator cv::Point3f() { return cv::Point3f(x,y,z); }
+BoardPoint::operator cv::Point2f() { return cv::Point2f(x,y); }
 BoardPoint::operator Eigen::Vector3d() { return Eigen::Vector3d(x,y,z); }
